[explorer/rest] select chain-specific route handlers
